fix(android): App baubar — 179 Kompilierfehler behoben, signierte Release-APK möglich (#18)
* feature: Betriebsgelände-Betrieb — Premises-Security, EDM-Lebenszyklus, Offline-Sync, Flotten-Gruppen, Interaktionsfeld

Kontext (verbindlich dokumentiert): EDM-verwaltetes Betriebsgelände, nur
eigene Geräte — Nutzung durch Dritte setzt EDM-Gerät + Betriebskenntnis voraus.

Neu implementiert:
- premises_security.py: passive Fremdgeräte-Erkennung (own/infra/unknown
  gegen eigene Register), premises_alert bei neuen Unbekannten (WS),
  passive Sensorberichte (Magnet/IR/RF, Stufe-2-Datenvertrag),
  Störungs-DETEKTION (Geräteschwund ≥3/≥50 %) — keine Gegenmaßnahmen
- edm.py: EDM-Lebenszyklus eigener Geräte (ENROLLED→PROVISIONED→
  QUARANTINED→RESET_PENDING→RESET), voll auditiert; Reset ausschließlich
  über Honeywell Provisioning Mode/OEMConfig (kein FRP-Bypass);
  illegale Übergänge (Direkt-Reset) → 403
- Sync-Queue: /api/v1/sync/queue|next|ack (Geräte-Scope, max. 200);
  sw.js: periodicsync/sync-Handler (Best-Effort); fleet.js:
  queueMeshObservation() + Registrierung
- Flotten-Gruppen: /api/v1/fleet/groups* (nur existierende Fahrzeuge);
  Positionshistorie /api/v1/fleet/{id}/history aus spatial_memory
- Interaktionsfeld (Accordion) im OSM-Dashboard: Fähigkeiten je Klasse,
  Gruppe/Quelle/Akku, Historie-Laden, Sichtbarkeit
- Metriken: 3dxagent_premises_unknown, 3dxagent_edm_devices

Bewusst NICHT enthalten (unverändert, dokumentiert): aktive Angriffswerkzeuge
(Deauth/Beacon-Spam/Rogue-AP/Reconnect-Flooding), Biometrie, taktisches
Vital-/Combat-Monitoring, Personensuche, FRP-Bypass für Fremdgeräte.

Qualität: 327 Tests grün (15 neue: Premises/EDM/API), Smoke-Test +
Fleet-E2E (16 Checks) gegen Live-Stack grün, 90 Routen.
Doku: docs/PREMISES_EDM.md

* fix(android): App baubar — 179 Kompilierfehler behoben, signierte Release-APK möglich

Warum es keine APK gab: unvollständig zusammengeführte Branches mit
Syntaxfehlern (3), doppeltem signingConfigs-Block und 179 Kotlin-Fehlern.

Behoben (Auswahl):
- Syntax: NetworkTraffic.kt (fehlender Klassen-/Funktionskopf),
  FloorPlanModels.kt (unvollständige SourceDescriptor-Zeile),
  AgentWebSocketClient.kt (2 fehlende Klammern)
- signingConfigs-Duplikat in app/build.gradle.kts entfernt
- AgentWebSocketClient: doppelte Feldblöcke dedupliziert, sendPayload
  ergänzt, AtomicBoolean-Zugriffe vereinheitlicht, Session-Konstanten
- MainActivity: EkfFusion/AgentApiClient-Importe, Gateway-Session-
  Verdrahtung (SecureCredentialStore + GatewaySessionManager),
  Alarm-UI-State (AlarmUiState), NotificationCompat, saveCurrentState
  (echte SpatialRecord-Persistenz), Fragment-Fragment entfernt
- UwbManager: ohne nicht-auflösbare androidx.uwb-Alpha-Bibliothek
  (Phasen-/NLOS-Kette bleibt; FiRa-Driver als dokumentierte Ergänzung)
- MedicalDriverFactory: ohne fehlende Treiber-Referenzen
  (RealMedicalMonitoringService als Fallback, dokumentiert)
- BluetoothAccessoryManager-Verdrahtung im BleTokenManager
- TacticalDashboardFragment: Layout fragment_tactical_dashboard.xml
  ergänzt, LinearLayoutManager-Import, safe-call-Zuweisungen
- Typ-/Sichtbarkeits-/Import-Fixes in 24 Dateien (aura, bridge, offline,
  pipeline, security, sensors, triangulation, wireless, device, maintenance)

Ergebnis: assembleRelease erfolgreich → signierte APK
(com.example.agent v4.5.0-BT-Accessories, minSdk 31, targetSdk 34)

---------

+++ android-app/app/src/main/java/com/example/agent/aura/RtiSolver.kt
@@ -130,7 +130,7 @@ class RtiSolver(
                 val dTx = distance3(c, link.tx)
                 val dRx = distance3(c, link.rx)
                 if (dTx + dRx < dLink + ellipseWidth) {
-                    val w = 1.0 / sqrt((dTx + dRx).toDouble()).toFloat()
+                    val w = 1.0f / sqrt((dTx + dRx).toDouble()).toFloat()
                     weights.add(i to w)
                     sum += w
                 }
@@ -229,7 +229,7 @@ class RtiSolver(
             val y = links[i].attenuationDb.toDouble()
             for ((v, w) in rows[i]) {
                 field[v] += w * y
-                weightSum[v] += w
+                weightSum[v] += w.toDouble()
             }
         }
         for (v in 0 until n) {
